+++ {{cookiecutter.repo_name}}/project/mods_available/icomoon/__init__.py
@@ -1,15 +1,11 @@
 """
 .. _Django Icomoon: https://github.com/sveetch/django-icomoon
 
-`Django Icomoon`_ display a gallery of all your defined Icomoon webfonts. It won't work with a webfont not generated on Icomoon site because it depends on a JSON manifest file.
+`Django Icomoon`_ help you to manage your webfonts with Icomoon service. It won't work with a webfont not generated on Icomoon site because it depends on a JSON manifest file (you could make it yourself but it's a little bit complicated).
 
-Default behavior is to search for the JSON manifest file at ``project/webapp_statics/fonts/selection.json``, so when you deploy an icomoon webfont you have to put there the ``selection.json`` file given into the download ZIP archive.
+This mod can handle many webfonts if you need, you just have to define them in the mod settings, at least one webfont is required.
 
-If you have more than one Icomoon webfont in your project, you can define more manifest to search for, read the documentation for more details.
+Once one or more webfonts are defined, `Django Icomoon`_ can help you to automatically deploy them in your project from downloaded Zip on Icomoon using a command line ``django-instance icomoon_deploy``.
 
-Once the JSON manifest installed, along the webfont files, you can reach the gallery at : ::
-
-    /icomoon/
-
-Default behavior is to required to be authenticated to access to the gallery.
+Also when deployed and the webfonts are loaded in your templates, you can visualize every icons from a gallery located at ``/icomoon/``.
 """